pyproject.toml
[project]
name = "iqm-quantum-school"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dimod>=0.12.21",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    #"iqm-client[qiskit]>=32.1.1,<33.0.0", # for day 1 to 3-1
    "iqm-client[qiskit]>=33.0.2", # for day 3-2 and 3-3
    "iqm-pulla>=12.0.2", # for day 3-2 and 3-3
    "iqm-pulse>=12.7.2",
    #"iqm-qaoa>=1.38.1", # for day 2
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pylatexenc>=2.10",
    "qrisp[iqm]>=0.7.17",
    "toolz>=1.1.0",
]
[qiskit]
